chore: update new code

+++ src/Domain/Common/IVersion.cs
@@ -2,5 +2,5 @@
 
 public interface IVersion
 {
-    int Version { get; set; }
+    long Version { get; set; }
 }
